refactor: ProtectedRoute 컴포넌트에서 인증 로직을 분리하여 protected-loader로 이동
