+++ src/router/components/ProtectedRoute.tsx
@@ -1,14 +1,4 @@
-import { useAuthStore } from '@/stores/auth';
-import { Outlet, redirect } from 'react-router-dom';
-
-export const protectedLoader = () => {
-  const accessToken = useAuthStore.getState().accessToken;
-
-  if (!accessToken) {
-    return redirect('/auth/login');
-  }
-  return null;
-};
+import { Outlet } from 'react-router-dom';
 
 export const ProtectedRoute = () => {
   return <Outlet />;
