+++ src/router/loader/protected-loader.ts
@@ -1,0 +1,11 @@
+import { useAuthStore } from '@/stores/auth';
+import { redirect } from 'react-router-dom';
+
+export const protectedLoader = () => {
+  const accessToken = useAuthStore.getState().accessToken;
+
+  if (!accessToken) {
+    return redirect('/auth/login');
+  }
+  return null;
+};
